sys/tools: add -s options for KCFI use
The KCFI requires all function signature match.

For .m files in KCFI, it calls kobj_error_methods in default.

Which means the signature will mismatched thus the KCFI will emit trap
in kernel.

Thus we proposed a option -s for makeobjops.awk to generate the correct
signature for each missing default function.

We do not enable this feature by default because it may enlarge the
kernel size.