atkbd: fix first keystroke force reset
Some i8042 falsely return KBD_ACK for ECHO command, which cuase the
keyboard echo test to fail.
Without passing the echo test, the keyboard is considered unconfigured.

Though it is an incorrect behavior, we still regard it as acceptable
to prevent the force reset (a step for configure a keyboard)
of the whole keyboard when the first key interrupt reached.

Co-Authored-By: Aymeric Wibo <obiwac@freebsd.org>
Reviewed By:    wulf obiwac emaste
Sponsored By:   FreeBSD Foundation
Differential Revision: https://reviews.freebsd.org/D50498