libusb: fix incorrect member name in bos_descriptor
The bNumDeviceCapabilities is bNumDeviceCaps when it is imported to
libusb upstream.
For a better compatibility, we should rename the member.
Also, for the backward compatibility, we use a macro to point
bNumDeviceCapabilities to bNumDeviceCaps
See: https://github.com/libusb/libusb/commit/02ebafc85d3f219842cbabaf78abc8100b6656e5

Reviewed By:    kevans
Sponsored By:   FreeBSD Foundation
Differential Revision: https://reviews.freebsd.org/D50740