usb: support SSP and SSPx2 xHCI
Currently, we only support default PSI id implemented xHCI. For other
xhci, we fallback to 3.0. Note that the total bandwidth of the device
is not so much impacted since all of the setup is done by xHCI
hardware id is compatbile for old xHCI driver.

This driver fix the addressing issue for SSP hub behind a SSP hub and
makes the usb driver detect the speed correctly.

Subscribers: imp

Differential Revision: https://reviews.freebsd.org/D52347

xhci: fix